refactored project folder structure
